Tests and re-organization
